Remove suggestion to delete "flintlock"
Xapian's `flintlock` is not a lock file whose presence indicates that
the lock is held, but rather a file which is locked using `fcntl()`
locking.  Such locks are automatically released by the OS when the
process holding them terminates, so there's no such thing as a "stale
lock", and the `flintlock` file should always exist.

The `flicklock` file was a typo and fixed back in 2006, before Xapian
1.0.0 - so the name only of historic interest now, but similarly it
should not be deleted.

The `db_lock` file was used by the old `quartz` backend and did work
by its presence indicating that the lock was held.  But `flint` was
the default backend from 1.0.0 and `quartz` was removed completely
in 2008, so I find it hard to believe that anyone will have such a
database.