Add support for gmime >= 3.0
The minimum required gmime version is now 2.6, though 2.4 may work
if you adjust the configure check (but gmime 2.4 is really old now,
and I'm not able to check with it).

Before this change it seems GMIME_ENABLE_RFC2047_WORKAROUNDS was
abused as a gmime version indicator - ChangeLog-dijon says:

> If GMIME_ENABLE_RFC2047_WORKAROUNDS is defined, assume gmime 2.4.

But this is no longer defined by gmime >= 3.0, and gmime < 2.4 seems
completely irrelevant now so I've stripped out these checks instead
of trying to adjust them to keep working.